refactor: update package structure and enhance contact synchronization endpoint
