--- chat/src/main/kotlin/chat/api/dto/ContactRequest.kt
@@ -1,7 +1,0 @@
-package chat.api.dto
-
-data class ContactRequest(
-    val firstName: String,
-    val lastName: String,
-    val phoneNumber: String
-)
