+++ chat/src/main/kotlin/net/thechance/chat/entity/Contact.kt
@@ -1,13 +1,14 @@
-package chat.entity
+package net.thechance.chat.entity
 
-import jakarta.persistence.Column
-import jakarta.persistence.Entity
-import jakarta.persistence.Id
-import jakarta.persistence.Table
+import jakarta.persistence.*
 import java.util.UUID
 
 @Entity
-@Table(name = "contacts", schema = "chat")
+@Table(
+    name = "contacts",
+    schema = "chat",
+    indexes = [Index(name = "idx_contacts_user_id", columnList = "userId")]
+)
 data class Contact(
     @Id @Column(columnDefinition = "uuid", updatable = false, nullable = false)
     val id: UUID = UUID.randomUUID(),
